requirements.txt
anyio==4.8.0
asgiref==3.8.1
build==1.2.2.post1
CacheControl==0.14.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cleo==2.1.0
crashtest==0.4.1
cryptography==44.0.1
distlib==0.3.9
Django==5.1.6
django-cors-headers==4.7.0
djangorestframework==3.15.2
dulwich==0.22.7
fastjsonschema==2.21.1
filelock==3.17.0
findpython==0.6.2
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
installer==0.7.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jeepney==0.8.0
keyring==25.6.0
more-itertools==10.6.0
msgpack==1.1.0
packaging==24.2
pbs-installer==2025.2.12
pkginfo==1.12.1.2
platformdirs==4.3.6
poetry==2.1.1
poetry-core==2.1.1
pycparser==2.22
pyproject_hooks==1.2.0
pytz==2025.1
RapidFuzz==3.12.1
# Editable install with no version control (realmate-challenge==0.1.0)
-e /home/paulo/Documents/patents/techinal-challlenges/realmate-challenge-main
requests==2.32.3
requests-toolbelt==1.0.0
SecretStorage==3.3.3
shellingham==1.5.4
sniffio==1.3.1
sqlparse==0.5.3
tomlkit==0.13.2
trove-classifiers==2025.2.18.16
urllib3==2.3.0
virtualenv==20.29.2
zstandard==0.23.0
